Break up Distribution.Package -- only classes and reexports remain
