+++ Cabal/Distribution/PackageDescription/Check.hs
@@ -45,10 +45,11 @@ import Distribution.Simple.BuildPaths (autogenPathsModuleName)
 import Distribution.Simple.BuildToolDepends
 import Distribution.Simple.CCompiler
 import Distribution.Types.ComponentRequestedSpec
+import Distribution.Types.CondTree
 import Distribution.Types.Dependency
 import Distribution.Types.ExeDependency
+import Distribution.Types.PackageName
 import Distribution.Types.UnqualComponentName
-import Distribution.Types.CondTree
 import Distribution.Simple.Utils hiding (findPackageDesc, notice)
 import Distribution.Version
 import Distribution.Package
