+++ Cabal/Distribution/Simple/GHCJS.hs
@@ -25,7 +25,6 @@ import Distribution.Simple.GHC.ImplInfo
 import qualified Distribution.Simple.GHC.Internal as Internal
 import Distribution.PackageDescription as PD
 import Distribution.InstalledPackageInfo
-import Distribution.Package
 import Distribution.Simple.PackageIndex ( InstalledPackageIndex )
 import qualified Distribution.Simple.PackageIndex as PackageIndex
 import Distribution.Simple.LocalBuildInfo
@@ -46,6 +45,7 @@ import Distribution.System
 import Distribution.Verbosity
 import Distribution.Utils.NubList
 import Distribution.Text
+import Distribution.Types.UnitId
 import Language.Haskell.Extension
 
 import qualified Data.Map as Map
